# recipe-server: Reduce the number of DB queries on API listing pages.