--- a/recipe-server/normandy/recipes/tests/api/v1/test_api.py
+++ b/recipe-server/normandy/recipes/tests/api/v1/test_api.py
@@ -18,6 +18,7 @@ from normandy.recipes.tests import (
     CountryFactory,
     LocaleFactory,
     RecipeFactory,
+    RecipeRevisionFactory,
     fake_sign,
 )
 
@@ -110,6 +111,17 @@ class TestActionAPI(object):
         assert res.data[0]['signature']['signature'] == a1.signature.signature
         assert res.data[1]['action']['name'] == a2.name
         assert res.data[1]['signature']['signature'] == a2.signature.signature
+
+    def test_it_makes_a_reasonable_number_of_db_queries(self, client):
+        # Naive versions of this view could easily make several queries
+        # *per action*, which is very slow. Make sure that isn't the case.
+        ActionFactory.create_batch(100)
+        queries = CaptureQueriesContext(connection)
+        with queries:
+            res = client.get('/api/v1/action/')
+            assert res.status_code == 200
+        # Anything under 100 isn't doing one query per recipe.
+        assert len(queries) < 100
 
 
 @pytest.mark.django_db
@@ -425,6 +437,17 @@ class TestRecipeAPI(object):
         assert len(res.data) == 1
         assert res.data[0]['recipe']['id'] == disabled_recipe.id
 
+    def test_signed_listing_makes_a_reasonable_number_of_db_queries(self, client):
+        # Naive versions of this view could easily make several queries
+        # *per recipe*, which is very slow. Make sure that isn't the case.
+        RecipeFactory.create_batch(100)
+        queries = CaptureQueriesContext(connection)
+        with queries:
+            res = client.get('/api/v1/recipe/signed/')
+            assert res.status_code == 200
+        # Anything under 100 isn't doing one query per recipe.
+        assert len(queries) < 100
+
     def test_list_sets_no_cookies(self, api_client):
         res = api_client.get('/api/v1/recipe/')
         assert res.status_code == 200
@@ -558,6 +581,17 @@ class TestRecipeAPI(object):
         r.refresh_from_db()
         assert list(r.channels.all()) == [c2]
 
+    def test_it_makes_a_reasonable_number_of_db_queries(self, client):
+        # Naive versions of this view could easily make several queries
+        # *per recipe*, which is very slow. Make sure that isn't the case.
+        RecipeFactory.create_batch(100)
+        queries = CaptureQueriesContext(connection)
+        with queries:
+            res = client.get('/api/v1/recipe/')
+            assert res.status_code == 200
+        # Anything under 100 isn't doing one query per recipe.
+        assert len(queries) < 100
+
 
 @pytest.mark.django_db
 class TestRecipeRevisionAPI(object):
@@ -585,6 +619,17 @@ class TestRecipeRevisionAPI(object):
         res = api_client.post(
             '/api/v1/recipe_revision/{}/request_approval/'.format(recipe.latest_revision.id))
         assert res.status_code == 400
+
+    def test_it_makes_a_reasonable_number_of_db_queries(self, client):
+        # Naive versions of this view could easily make several queries
+        # *per revision*, which is very slow. Make sure that isn't the case.
+        RecipeRevisionFactory.create_batch(100)
+        queries = CaptureQueriesContext(connection)
+        with queries:
+            res = client.get('/api/v1/recipe_revision/')
+            assert res.status_code == 200
+        # Anything under 100 isn't doing one query per recipe.
+        assert len(queries) < 100
 
 
 @pytest.mark.django_db
@@ -658,6 +703,17 @@ class TestApprovalRequestAPI(object):
 
         with pytest.raises(ApprovalRequest.DoesNotExist):
             ApprovalRequest.objects.get(pk=a.pk)
+
+    def test_it_makes_a_reasonable_number_of_db_queries(self, client):
+        # Naive versions of this view could easily make several queries
+        # *per action*, which is very slow. Make sure that isn't the case.
+        ApprovalRequestFactory.create_batch(100)
+        queries = CaptureQueriesContext(connection)
+        with queries:
+            res = client.get('/api/v1/approval_request/')
+            assert res.status_code == 200
+        # Anything under 100 isn't doing one query per recipe.
+        assert len(queries) < 100
 
 
 @pytest.mark.django_db