--- a/recipe-server/normandy/recipes/tests/api/v2/test_api.py
+++ b/recipe-server/normandy/recipes/tests/api/v2/test_api.py
@@ -75,20 +75,6 @@ class TestActionAPI(object):
         assert res.status_code == 200
         assert res.client.cookies == {}
 
-    def test_it_makes_a_reasonable_number_of_db_queries(self, client, settings):
-        # Naive versions of this view could easily make several queries
-        # *per action*, which is very slow. Make sure that isn't the case.
-        ActionFactory.create_batch(100)
-        queries = CaptureQueriesContext(connection)
-        with queries:
-            res = client.get('/api/v2/action/')
-            assert res.status_code == 200
-
-        # Pagination naturally makes one query per item in the page. Anything
-        # under `page_size * 2` isn't doing any additional queries per recipe.
-        page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
-        assert len(queries) < page_size * 2
-
 
 @pytest.mark.django_db
 class TestRecipeAPI(object):
@@ -455,20 +441,6 @@ class TestRecipeAPI(object):
         r.refresh_from_db()
         assert list(r.channels.all()) == [c2]
 
-    def test_it_makes_a_reasonable_number_of_db_queries(self, client, settings):
-        # Naive versions of this view could easily make several queries
-        # *per action*, which is very slow. Make sure that isn't the case.
-        RecipeFactory.create_batch(100)
-        queries = CaptureQueriesContext(connection)
-        with queries:
-            res = client.get('/api/v2/recipe/')
-            assert res.status_code == 200
-
-        # Pagination naturally makes one query per item in the page. Anything
-        # under `page_size * 2` isn't doing any additional queries per recipe.
-        page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
-        assert len(queries) < page_size * 2
-
 
 @pytest.mark.django_db
 class TestRecipeRevisionAPI(object):
@@ -496,20 +468,6 @@ class TestRecipeRevisionAPI(object):
         res = api_client.post(
             '/api/v2/recipe_revision/{}/request_approval/'.format(recipe.latest_revision.id))
         assert res.status_code == 400
-
-    def test_it_makes_a_reasonable_number_of_db_queries(self, client, settings):
-        # Naive versions of this view could easily make several queries
-        # *per action*, which is very slow. Make sure that isn't the case.
-        RecipeRevisionFactory.create_batch(100)
-        queries = CaptureQueriesContext(connection)
-        with queries:
-            res = client.get('/api/v2/recipe_revision/')
-            assert res.status_code == 200
-
-        # Pagination naturally makes one query per item in the page. Anything
-        # under `page_size * 2` isn't doing any additional queries per recipe.
-        page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
-        assert len(queries) < page_size * 2
 
 
 @pytest.mark.django_db
@@ -583,20 +541,6 @@ class TestApprovalRequestAPI(object):
 
         with pytest.raises(ApprovalRequest.DoesNotExist):
             ApprovalRequest.objects.get(pk=a.pk)
-
-    def test_it_makes_a_reasonable_number_of_db_queries(self, client, settings):
-        # Naive versions of this view could easily make several queries
-        # *per action*, which is very slow. Make sure that isn't the case.
-        ApprovalRequestFactory.create_batch(100)
-        queries = CaptureQueriesContext(connection)
-        with queries:
-            res = client.get('/api/v2/approval_request/')
-            assert res.status_code == 200
-
-        # Pagination naturally makes one query per item in the page. Anything
-        # under `page_size * 2` isn't doing any additional queries per recipe.
-        page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
-        assert len(queries) < page_size * 2
 
 
 @pytest.mark.django_db
@@ -782,3 +726,25 @@ class TestApprovalFlow(object):
 
         # The API should still have correct signatures
         self.verify_signatures(api_client, expected_count=1)
+
+
+@pytest.mark.django_db
+@pytest.mark.parametrize("endpoint,Factory", [
+    ('/api/v2/action/', ActionFactory),
+    ('/api/v2/recipe/', RecipeFactory),
+    ('/api/v2/recipe_revision/', RecipeRevisionFactory),
+    ('/api/v2/approval_request/', ApprovalRequestFactory),
+])
+def test_apis_makes_a_reasonable_number_of_db_queries(endpoint, Factory, client, settings):
+    # Naive versions of this view could easily make several queries
+    # per item, which is very slow. Make sure that isn't the case.
+    Factory.create_batch(100)
+    queries = CaptureQueriesContext(connection)
+    with queries:
+        res = client.get(endpoint)
+        assert res.status_code == 200
+
+    # Pagination naturally makes one query per item in the page. Anything
+    # under `page_size * 2` isn't doing any additional queries per recipe.
+    page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
+    assert len(queries) < page_size * 2