@Osmose 2017-08-30
RE: de-duplicating test logic, perhaps we make this a standalone test generator instead of adding it to the per-view tests? Seems like the URL and the factory are the main things changing between tests.